+++ packages/perseus-editor/src/components/locked-figures-section.tsx
@@ -10,7 +10,7 @@ import * as React from "react";
 
 import LockedFigureSelect from "./locked-figure-select";
 import LockedFigureSettings from "./locked-figure-settings";
-import {getDefaultFigureForFigureType} from "./util";
+import {getDefaultFigureForType} from "./util";
 
 import type {Props as InteractiveGraphEditorProps} from "../widgets/interactive-graph-editor";
 import type {LockedFigure, LockedFigureType} from "@khanacademy/perseus";
@@ -29,7 +29,7 @@ const LockedFiguresSection = (props: Props) => {
         const newProps = {
             lockedFigures: [
                 ...lockedFigures,
-                getDefaultFigureForFigureType(newFigure),
+                getDefaultFigureForType(newFigure),
             ],
         };
         onChange(newProps);
@@ -45,38 +45,24 @@ const LockedFiguresSection = (props: Props) => {
         });
     }
 
-    function changeCoord(index: number, coord: [number, number]) {
+    function changeProps(
+        index: number,
+        // Omit the type from the figure props so it doesn't think
+        // we're trying to pass in the props for the wrong type.
+        figureProps: Omit<Partial<LockedFigure>, "type">,
+    ) {
         const lockedFigures = figures || [];
-        const newProps = {
+        const newFigures = {
             lockedFigures: [
                 ...lockedFigures.slice(0, index),
                 {
                     ...lockedFigures[index],
-                    coord,
+                    ...figureProps,
                 },
                 ...lockedFigures.slice(index + 1),
             ],
         };
-        onChange(newProps);
-    }
-
-    function changeColor(index: number, colorName: string) {
-        const lockedFigures = figures || [];
-        const newProps = {
-            lockedFigures: [
-                ...lockedFigures.slice(0, index),
-                {
-                    ...lockedFigures[index],
-                    style: {
-                        ...lockedFigures[index].style,
-                        fill: colorName,
-                        stroke: colorName,
-                    },
-                },
-                ...lockedFigures.slice(index + 1),
-            ],
-        };
-        onChange(newProps);
+        onChange(newFigures);
     }
 
     return (
@@ -85,8 +71,7 @@ const LockedFiguresSection = (props: Props) => {
                 <LockedFigureSettings
                     key={`${uniqueId}-locked-${figure}-${index}`}
                     {...figure}
-                    onChangeColor={(color) => changeColor(index, color)}
-                    onChangeCoord={(coord) => changeCoord(index, coord)}
+                    onChangeProps={(newProps) => changeProps(index, newProps)}
                     onRemove={() => removeLockedFigure(index)}
                 />
             ))}