+++ packages/perseus/src/index.ts
@@ -64,6 +64,7 @@ export {
     PerseusExpressionAnswerFormConsidered,
     plotterPlotTypes,
     ItemExtras,
+    lockedFigureColors,
 } from "./perseus-types";
 export {traverse} from "./traversal";
 export {isItemRenderableByVersion} from "./renderability";
@@ -157,8 +158,10 @@ export type {
 export type {ParsedValue} from "./util";
 export type {
     LockedFigure,
+    LockedFigureColor,
     LockedFigureType,
-    LockedPoint,
+    LockedPointType,
+    LockedLineType,
     PerseusGraphType,
     PerseusAnswerArea,
     PerseusExpressionWidgetOptions,